Change udf_dirstream to use more specific integer types, chiefly off_t for offsets and size_t for sizes.
OK natano@.

+++ sys/isofs/udf/udf.h
@@ -93,11 +93,11 @@ struct udf_dirstream {
 	struct buf	*bp;
 	uint8_t		*data;
 	uint8_t		*buf;
-	int		fsize;
-	int		off;
-	int		this_off;
-	int		offset;
-	int		size;
+	size_t		fsize;
+	off_t		off;
+	off_t		this_off;
+	off_t		offset;
+	size_t		size;
 	int		error;
 	int		fid_fragment;
 };
